+++ packages/roosterjs-editor-plugins/lib/plugins/TableCellSelection/features/DeleteTableContents.ts
@@ -1,0 +1,39 @@
+import { safeInstanceOf } from 'roosterjs-editor-dom';
+import {
+    BuildInEditFeature,
+    IEditor,
+    Keys,
+    PluginKeyboardEvent,
+    SelectionRangeTypes,
+} from 'roosterjs-editor-types';
+
+/**
+ * @internal
+ * Feature that when Backspace is pressed and there is Table Selection, delete the contents inside of the selection
+ */
+export const DeleteTableContents: BuildInEditFeature<PluginKeyboardEvent> = {
+    keys: [Keys.DELETE, Keys.BACKSPACE],
+    shouldHandleEvent: (_, editor: IEditor) => {
+        const selection = editor.getSelectionRangeEx();
+        return selection.type == SelectionRangeTypes.TableSelection;
+    },
+    handleEvent: (_, editor) => {
+        const selection = editor.getSelectionRangeEx();
+        if (selection.type == SelectionRangeTypes.TableSelection) {
+            editor.addUndoSnapshot(() => {
+                editor.getSelectedRegions().forEach(region => {
+                    if (safeInstanceOf(region.rootNode, 'HTMLTableCellElement')) {
+                        deleteNodeContents(region.rootNode, editor);
+                    }
+                });
+            });
+        }
+    },
+};
+
+function deleteNodeContents(element: HTMLElement, editor: IEditor) {
+    const range = new Range();
+    range.selectNodeContents(element);
+    range.deleteContents();
+    element.appendChild(editor.getDocument().createElement('br'));
+}
